engine/requirements.txt
manim
ffmpeg-python
fastapi
uvicorn
websockets
Pillow
pytest
black
ruff
langchain
google-generativeai
openai
